Add support for variable support to "SELECT SKIP <ROWCOUNT> FIRST <ROWCOUNT>..." construct
The grammar for the construct in informix [1] mentions the possibility, that <ROWCOUNT> can be either
an integer or a host variable or local SPL variable storing the value of max. The case for plain integers and
jdbc variables is covered by the first commit While this commit adds support for constructs using SPL
variables. SPL variables must follow identifier rules [2][3].

[1] http://www-01.ibm.com/support/knowledgecenter/SSGU8G_12.1.0/com.ibm.sqls.doc/ids_sqs_0156.htm
[2] http://www-01.ibm.com/support/knowledgecenter/SSGU8G_12.1.0/com.ibm.sqls.doc/ids_sqs_1306.htm?lang=de
[3] http://www-01.ibm.com/support/knowledgecenter/SSGU8G_12.1.0/com.ibm.sqls.doc/ids_sqs_1660.htm%23ids_sqs_1660?lang=de