Add ability to support "NOT LIKE ..." expressions
The parser is able to parse expressions such as "a NOT LIKE '%pattern%'", but is not able to parse expressions where the not is before the entire expression. For example: "NOT a LIKE '%pattern%'.

When parsing the latter, the error is:
Caused by: net.sf.jsqlparser.parser.ParseException: Encountered " "LIKE" "LIKE "" at line 1, column 32.
Was expecting one of: ...

The reason this is important is both because these syntaxes are both valid, and also because the deparser uses the second method.
Therefore, if you parse a query with the first type of expression, then deparse it and parse again, you'll get the same error.