Allow single quotes around aliases in SELECT clause
MySQL supports single quotes around aliases (in addition to double quotes and backticks). They are only allowed in aliases in the SELECT clause. In other cases, only backticks are supported.
https://dev.mysql.com/doc/refman/5.7/en/identifiers.html
From the documentation:
"The identifier quote character is the backtick (`)".
"In the select list of a query, a quoted column alias can be specified using identifier or string quoting characters".