Guard Values against null/empty values
The classes modified by this commit are `DoubleValue`, `LongValue`, and
`TimeValue`. Both `null` and empty strings provided to their
constructors fail, but they provide very different error messages
(NullPointerException and StringIndexOutOfBoundsException), which is
neither sensible nor helpful in debugging.
This commit adds a guard to throw `IllegalArgumentException` for both
cases in order to improve coherency and usefulness of the error
messages.