Fix[2306] - Adds support for Trino UDF (#2307)
* Fix[2306] - Adds support for Trino UDF
Trino SQL allows you to specify User Defined functions in the WITH clause ahead of a statement. This adds support for this type of statement.

* Fix[2306] - Adds support for Trino UDF
Trino SQL allows you to specify User Defined functions in the WITH clause ahead of a statement. This adds support for this type of statement.

---------

Co-authored-by: David Hayes <dhayes@adobe.com>